refactor: dedupe hex-encoding loops into shared hash module
Four files hand-rolled the same `write!("{:02x}")` hex loop: idl.rs and
run_state.rs (`hex_encode`), setup.rs (`sha256_hex`), and wallet_support.rs
(inline `0x`-prefixed tx hash). Consolidate into `crate::hash` with
`hex_encode`/`sha256_hex` and call it from each site. Behavior unchanged.
