+++ src/commands/run_state.rs
@@ -25,6 +25,7 @@ use crate::commands::wallet_support::{
     default_sequencer_http_url_for_project, load_wallet_runtime,
 };
 use crate::constants::FRAMEWORK_KIND_LEZ_FRAMEWORK;
+use crate::hash::hex_encode;
 use crate::model::Project;
 use crate::state::read_localnet_state;
 use crate::DynResult;
@@ -226,15 +227,6 @@ fn state_path(project_root: &Path) -> std::path::PathBuf {
     project_root.join(RUN_DEPLOY_STATE_REL)
 }
 
-fn hex_encode(bytes: &[u8]) -> String {
-    use std::fmt::Write;
-    let mut s = String::with_capacity(bytes.len() * 2);
-    for b in bytes {
-        let _ = write!(s, "{b:02x}");
-    }
-    s
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -302,11 +294,6 @@ mod tests {
         let prior_empty = state_with(&[], Some(42));
         assert!(!deploy_can_be_skipped(&empty, Some(42), &prior_with));
         assert!(!deploy_can_be_skipped(&nonempty, Some(42), &prior_empty));
-    }
-
-    #[test]
-    fn hex_encode_is_lowercase_padded() {
-        assert_eq!(hex_encode(&[0x0a, 0xff, 0x00]), "0aff00");
     }
 
     #[test]
