+++ src/lib.rs
@@ -8,6 +8,7 @@ pub(crate) mod config;
 pub(crate) mod constants;
 pub(crate) mod doctor_checks;
 pub(crate) mod error;
+pub(crate) mod hash;
 pub(crate) mod migrate;
 pub(crate) mod model;
 pub(crate) mod process;
